backend/pyproject.toml
[project]
name = "analytics-chatbot-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "google-auth>=2.29.0",
  "google-cloud-bigquery>=3.21.0",
  "google-cloud-aiplatform>=1.70.0",
  "google-adk>=1.0.0",
  "google-genai>=1.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "httpx>=0.27.0",
  "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "auth*", "jobs*", "middleware*", "models*", "services*", "tools*"]
exclude = ["tests*"]
